Fix api

+++ backend/Data/Exercise.cs
@@ -31,7 +31,7 @@ namespace backend.Data
         [Required]
         public string HintCode { get; set; }
 
-        public float TimeLimit { get; set; }
+        public int TimeLimit { get; set; }
 
         public int Score { get; set; }
     }
