+++ backend/RequestModel/ExerciseModel.cs
@@ -17,6 +17,6 @@ namespace backend.RequestModel
 
         public string HintCode { get; set; }
 
-        public float TimeLimit { get; set; }
+        public int TimeLimit { get; set; }
     }
 }
